Switch to nuzzle.edn config format
This patch overhauls the way Nuzzle is configured. Instead of passing a
config map to all of Nuzzle's API functions, now that config map is also
stored on disk. The config map and the site-data have now merged into
the same EDN file which must be located at $CWD/nuzzle.edn.

This patch also adds the ability to override configuration via keyword
arguments for all of Nuzzle's API functions.

This change is awesome for a few reasons:
1. API function calls are simpler
2. Filesystem based hot-reloading for full config, not just site-data
3. Makes setting up a user.clj REPL entrypoint less necessary
4. Opens the door for using Nuzzle's API directory from deps.edn with
   :exec-fn in a super elegant way
5. Config is centralized