Avoid double loading config by passing it through request map
Allows the site index building to access the incoming request map. This
enables Nuzzle to only loading the config if it was already loaded by a
previous middleware, like wrap-overlay-dir which adds it the request
map. I had to wrap stasis.core/serve-pages because stasis does not give
the get-pages function access to the incoming request map.