Change :index and :tags from vectors to sets
This patch starts an initiative to migrate data types from vectors to
sets when sets are more fitting. I've been using vectors for every
collection in the configuration but any collection that is concerned
with membership and not order should be a set without exception. Not
only does that more clearly express the intent of the data, but it also
makes testing easier by avoiding random reordering test failures.