Avoid logging overlay-dir multiple times per page load
This patch fixes a bug where the overlay-dir would be logged potentially
many times on one page load because every static asset request from the
overlay-dir would also trigger a logging. This actually brought to my
attention that Nuzzle now reloads the config for every single request,
including static assets. It's not ideal, but this way Nuzzle can detect
when the overlay-dir changes in the user's nuzzle.edn.