Add support for line numbers in md syntax highlighting
This patch lets the user add line numbers to their markdown code syntax
highlighting with both Chroma and Pygments. The Pygments support is a
little janky because you have to wrap the output in a pre and div tag in
order to get line numbers. I submitted a PR upstream to fix this.

https://github.com/pygments/pygments/issues/2127