Refactor shortcode functions
Now shortcode functions take the whole hiccup vector as their argument
which gives them access to the tags within. Also the shortcode function
calls are explicitly dispatched off of a case statement now so static
analysis can catch errors easier. Also the hiccup attribute map is
ensured to be present even if empty when handed off to the shortcode
function.