Rename markdown and render-markdown to content and render-content
This is a basically a revert. I decided to switch back to content and
render-content as a compromise. I want Nuzzle to focus on making
Markdown great, but I don't want to exclude anyone who hates Markdown.
The plan now is to support HTML content as well so people can convert
whatever markup syntax they use to HTML and still get some of the
benefits of Nuzzle offers.