Add Malli decoding step when loading config
Now the incoming config will be decoded according to the schema and
provided transformations. This way we can read in the :modified date as
a string, decode into a java.time.LocalDate, and validate it. Using the
`#inst` tag is problematic because Clojure 1.11.1 still parses this into
a java.util.Date which is not at all ideal (timezone issues mostly).