Add support for generating sitemaps
With this patch, Nuzzle can now generate sitemap XML files which are the
standard way of informing web crawlers what your site looks like. The
sitemap generation requires a rendered site-index which documents which
pages will get rendered to disk. The site-data alone is not enough
because the webpage will not get exported is the webpage render function
returns nil. The sitemap includes the last modified date of a webpage if
the :modified key is present.