Add nuzzle.publish-test ns with publishing tests (#24)
The functions in nuzzle.publish are now tested. The
nuzzle.utils/diff-dirs function is being utilized to detect changes in
build outputs after code changes in the nuzzle.publish-test/publish-site
test. When a change is detected, the test will print out diffs (using
the diff shell utility) so diffs can be easily examined in CI failures.