Pull out :shortcode-fns to config top-level, rename to :nuzzle/custom-elements (#37)
The shortcode-fns setting is not well documented right now, but it
basically provides a way to define custom HTML tags with functions that
transform that element when it represented as Hiccup. This config
setting was previously located inside :markdown-opts, but now it's at
the top-level because it will eventually transcend Markdown usage to be
used with raw HTML as well.

I think custom-elements is a better name because it actually reflects
HTML custom elements in the browser quite well, it's just that these
custom elements are being rendered statically server side.

Adding nuzzle namespace to keyword to go along with #6.