Add clojure.spec rough draft to schemas
Malli is coming up short in regards to being able to fully encapsulate
the complexity of the Nuzzle config. There is no good way to define a
map where some values are validated based on the type of the keys.
clojure.spec is actually much better in this regard so I'll probably be
switching over soon.