Remove :site-data, merge into top-level config (#5)
Previously all page entries in the config map were stored inside a set
under the key :site-data. This change takes all the page entries in that
site-data set and lifts them to the top level of the config map with a
key of the page's :id value. This simplifies the config map greatly.

A ramification of this change is that the name get-site-data no longer
makes sense. This function has been changed to get-config and can be
used to get any value from the config map, making it even more useful.