Allow time vals to be date, datetime, or zoned datetime (#102)
Previously time values in the config could only be dates. Now time
values can be dates, datetimes, or zoned datetimes for more flexibility.
Zoned datetimes are the best because they are the only one that
explicitly refers to a moment in time irrespective of the user's current
time zone. The only time value in the config is :nuzzle/updated
currently but more will be added soon.

The tests involving these time values must always used zoned datetimes
because if not they will fail when run in other time zones.