Add author registry and page authors to config (#98)
This change introduces a new feature where pages can be tagged with a
`:nuzzle/author` key with a keyword value like `:laura` that represents
an author. All author keywords must be present in the top-level
`:nuzzle/author-registry` map which contains information about these
authors.

This feature is part of the migration to the Atom feed integration. Atom
defines authors for both the feed and individual pages. The author
registry will be used for both of these purposes.