Replace RSS feed with Atom feed generation (#95)
Previously Nuzzle generated an RSS feed based on the
`:nuzzle/rss-channel` top-level config value and the `:nuzzle/rss?` page
entry values. Now Nuzzle generates an Atom feed instead based on the
`:nuzzle/atom-feed` top-level config value and the `:nuzzle/feed?` page
entry values.

Normally Atom feeds cannot be deterministically generated because the
feed element requires an updated timestamp which should normally be the
time when the program runs. I've added a :deterministic? option to the
Atom feed generation to omit the feed's updated tag, allowing the tests
to get consistent output. In the future I may want to change this to
injecting an inst into the updated value because technically the test
Atom feeds are invalid.