Make laziness toggle for create-render-content-fn (#118)
For publishing the static site it's ideal to render the page content
eagerly and return a static value from `render-content` so the content
files never are processed more than once. For the server it's ideal for
`render-content` to lazily process the content files because only one
page needs to be rendered per request and usually not all the content
files will need to be processed.

This change adds an options map to `create-render-content-fn` with a key
called :lazy-render? that toggles whether the `render-content` functions
should be lazy or not. This way we get the best of both worlds for both
publishing and using the development server.