Add :nuzzle/ignore-pages config option (#130)
This change adds the :nuzzle/ignore-pages option which is a set of page
keys that must not be in the config after the transformation process.
This allows the user to ignore pages that Nuzzle automatically
generates.

Previously this was done by returning nil from the user's render-page
function. Now this method is abandoned. Returning nil will result in
errors when trying to export and blank pages when using the development
server.