Require literal config map, remove nuzzle.edn (#133)
Stop using nuzzle.edn to hold config value and start requiring that the
config is passed directly into the functions in `nuzzle.api`. This
allows the user to directly embed functions in the config map and allows
the user to pre-process the config map with functions provided by
Nuzzle in the future.

This is a big step in my vision towards a Nuzzle interface that puts
more control into the hands of users and doesn't do a lot of magic
behind the scenes.