Change config map to pages map with just page info
Previously we called the most important parameter to all the API
functions the "config". Now it is called "pages" because there are no
more values in the map besides page values (vector of keywords keys
representing URL, map values with metadata about the page). This is a
significant shift in Nuzzle's API.